Move PreviewCleaner to a separate class.
